+++ CHANGELOG.md
@@ -6,8 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## 11.6.0
 ### Added
-- Adding options for defining `library_version` and `platform_version` 
-- Adding library version and platform version to composer.json and package.json
+- Added options for defining `library_version` and `platform_version` 
+- Added library version and platform version to composer.json and package.json
 
 ## 11.5.0
 ### Changed